deleted # Further Exploration
# The order in which you perform tests for a leap year calculation is important.

# my solution:
# def leap_year?(year)
#   (year % 4 == 0) && (year % 100 != 0) || (year % 400 == 0)
# end

# For what years will leap_year? fail if you rewrite it as:
# 2400, 240000, 2000, and 400

# LS Solution:
def leap_year?(year)
  if year % 400 == 0
    true
  elsif year % 100 == 0
    false
  else
    year % 4 == 0
  end
end

Can you rewrite leap_year? to perform its tests in the opposite order of the above solution? That is, test whether the year is divisible by 4 first, then, if necessary, test whether it is divisible by 100, and finally, if necessary, test whether it is divisible by 400. Is this solution simpler or more complex than the original solution?