deleted Explanation of Initial Given Syntax:
Syntax:
a = %w(a b c d e)

Explanation:
a is a variable.
= means assignment.
%w is used to create an array of string literals, where each string is separated by whitespace.
(a b c d e) are string elements.
this means:
the variable a is assigned to the (a b c d e) array of string literals, where each string is separated by whitespace.
the result of which is:
[a, b, c, d, e]